refactor: helper function to replace repetitive code blocks, split auth_api into smaller files
