+++ landing-page/scripts/inference_api.py
@@ -1,13 +1,13 @@
-from pathlib import Path
 from typing import Optional
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel
 from fastapi.responses import StreamingResponse
 import json
 import uuid as _uuid
 import io
 
-from auth_api import get_db_conn, get_current_user, release_db_conn, get_model_file_path, require_project_owner
+from auth_api import get_db_conn, get_current_user, release_db_conn, db_cursor, require_project_owner
+from models_api import get_model_file_path
 from text_api import stream_text_finding
 
 router = APIRouter()
@@ -29,12 +29,8 @@ async def run_predict(
     body: PredictBody,
     user=Depends(get_current_user),
 ):
-    con = get_db_conn()
-    cur = con.cursor()
-    try:
+    with db_cursor() as (con, cur):
         require_project_owner(cur, project_id, user["id"])
-    finally:
-        cur.close(); release_db_conn(con)
 
     model_id = body.model_id
     image_ids = body.image_ids
@@ -152,3 +148,32 @@ async def run_predict(
         media_type="text/event-stream",
         headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
     )
+
+
+@router.delete("/projects/{project_id}/annotations/{annotation_id}")
+def delete_annotation(project_id: int, annotation_id: str, user=Depends(get_current_user)):
+    with db_cursor() as (con, cur):
+        require_project_owner(cur, project_id, user["id"])
+        cur.execute("DELETE FROM annotations WHERE id=%s AND project_id=%s", (annotation_id, project_id))
+        con.commit()
+        return {"ok": True}
+
+
+@router.get("/projects/{project_id}/annotations/{annotation_id}")
+async def get_annotation_txt(
+    project_id: int,
+    annotation_id: str,
+    user=Depends(get_current_user),
+):
+    with db_cursor() as (con, cur):
+        cur.execute(
+            "SELECT a.yolo_txt, a.image_name"
+            " FROM annotations a"
+            " JOIN projects p ON p.id = a.project_id"
+            " WHERE a.id = %s AND a.project_id = %s AND p.user_id = %s",
+            (annotation_id, project_id, user["id"]),
+        )
+        row = cur.fetchone()
+    if not row:
+        raise HTTPException(status_code=404)
+    return {"yoloTxt": row[0], "imageName": row[1]}
